its time to commit
