--- Game1/Engine/Pathfinding2/BinaryTree.cs
@@ -1,17 +1,0 @@
-﻿namespace Game1.Engine.Pathfinding2
-{
-    public class BinaryTree : IBinaryTree
-    {
-        public IBinaryTreeNode Root { get; set; }
-
-        public BinaryTree()
-        {
-            Root = null;
-        }
-
-        public void Clear()
-        {
-            Root = null;
-        }
-    }
-}
