--- Game1/Engine/Pathfinding2/IBinaryTree.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Game1.Engine.Pathfinding2
-{
-    public interface IBinaryTree
-    {
-        IBinaryTreeNode Root { get; set; }
-        void Clear();
-    }
-}
