--- Game1/Engine/Pathfinding2/IBinaryTreeNode.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Game1.Engine.Pathfinding2
-{
-    public interface IBinaryTreeNode
-    {
-        IBinaryTreeNode Left { get; set; }
-        IBinaryTreeNode Right { get; set; }
-    }
-}
