+++ Game1/EngineCode.csproj
@@ -53,26 +53,16 @@
     <Compile Include="Engine\Input\Controller\iControllerObserverable.cs" />
     <Compile Include="Engine\Input\Keyboard\KeyboardInput.cs" />
     <Compile Include="Engine\Managers\iManager.cs" />
-    <Compile Include="Engine\Pathfinding2\BinaryTree.cs" />
-    <Compile Include="Engine\Pathfinding2\BinaryTreeNode.cs" />
-    <Compile Include="Engine\Pathfinding2\Grid.cs" />
-    <Compile Include="Engine\Pathfinding2\IBinaryTree.cs" />
-    <Compile Include="Engine\Pathfinding2\IBinaryTreeNode.cs" />
-    <Compile Include="Engine\Pathfinding2\IGrid.cs" />
-    <Compile Include="Engine\Pathfinding2\INode.cs" />
-    <Compile Include="Engine\Pathfinding2\IPathFinding.cs" />
-    <Compile Include="Engine\Pathfinding2\Node.cs" />
-    <Compile Include="Engine\Pathfinding2\PathFinding.cs" />
-    <Compile Include="Engine\PathFinding\BinaryTree.cs" />
-    <Compile Include="Engine\PathFinding\BinaryTreeNode.cs" />
-    <Compile Include="Engine\PathFinding\BreadthFirstSearch.cs" />
-    <Compile Include="Engine\PathFinding\DepthFirstSearch.cs" />
-    <Compile Include="Engine\PathFinding\IBinaryTree.cs" />
-    <Compile Include="Engine\PathFinding\IBinaryTreeNode.cs" />
-    <Compile Include="Engine\PathFinding\IBreadthFirstSearch.cs" />
-    <Compile Include="Engine\PathFinding\IDepthFirstSearch.cs" />
-    <Compile Include="Engine\PathFinding\IPathFindingNode.cs" />
-    <Compile Include="Engine\PathFinding\PathFindingNode.cs" />
+    <Compile Include="Engine\Pathfinding\BinaryTree.cs" />
+    <Compile Include="Engine\Pathfinding\BinaryTreeNode.cs" />
+    <Compile Include="Engine\Pathfinding\Grid.cs" />
+    <Compile Include="Engine\Pathfinding\IBinaryTree.cs" />
+    <Compile Include="Engine\Pathfinding\IBinaryTreeNode.cs" />
+    <Compile Include="Engine\Pathfinding\IGrid.cs" />
+    <Compile Include="Engine\Pathfinding\INode.cs" />
+    <Compile Include="Engine\Pathfinding\IPathFinding.cs" />
+    <Compile Include="Engine\Pathfinding\Node.cs" />
+    <Compile Include="Engine\Pathfinding\PathFinding.cs" />
     <Compile Include="Engine\Render\Renderer.cs" />
     <Compile Include="Engine\Collision\CollisionManager.cs" />
     <Compile Include="Engine\Collision\CollisionMap.cs" />
